requirements.txt
streamlit
google-api-python-client
google-auth
google-generativeai
PyPDF2
python-dotenv
google-cloud-firestore